refactor: 매칭 취소, 찔러보기 취소 버튼 분리
