+++ app/my-page/_components/MatchCancelButton.tsx
@@ -1,0 +1,46 @@
+import { Button } from "@/app/_common/shadcn/ui/button";
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/app/_common/shadcn/ui/alert-dialog";
+
+import { useMutationCancelMatch } from "../_hooks/useMutationCancelMatch";
+
+interface MatchCancelButtonProps {
+  id: number;
+}
+
+function MatchCancelButton({ id }: MatchCancelButtonProps) {
+  const { mutate } = useMutationCancelMatch();
+
+  return (
+    <AlertDialog>
+      <AlertDialogTrigger>
+        <Button>취소</Button>
+      </AlertDialogTrigger>
+      <AlertDialogContent>
+        <AlertDialogHeader>
+          <AlertDialogTitle>정말 매치를 취소하시겠습니까?</AlertDialogTitle>
+          <AlertDialogDescription>
+            이 작업은 되돌릴 수 없습니다.
+          </AlertDialogDescription>
+        </AlertDialogHeader>
+        <AlertDialogFooter>
+          <AlertDialogCancel>아니요</AlertDialogCancel>
+          <AlertDialogAction onClick={() => mutate(id)}>
+            매치 취소하기
+          </AlertDialogAction>
+        </AlertDialogFooter>
+      </AlertDialogContent>
+    </AlertDialog>
+  );
+}
+
+export default MatchCancelButton;
